✨ Feat : 사용자 프로필 이미지 저장 API

+++ business/src/main/java/com/mzbr/business/member/controller/MemberController.java
@@ -74,7 +74,7 @@ public class MemberController {
 	public ResponseEntity<Void> changeProfileImage(@RequestParam("profileImage") MultipartFile image,
 		@AuthenticationPrincipal UserDetails userDetails) {
 
-		memberService.changeProfileImage(image, Integer.parseInt(userDetails.getUsername()));
+		memberService.changeProfileImage(image, Long.parseLong(userDetails.getUsername()));
 
 		return ResponseEntity.ok().build();
 	}
